Sort candidates to solve by the frequency of appearance
For a given cell, process the candidates in order of "likelihood."
"Likelihood" is highest if that candidate appears as a possibility in
the most cells. This shaves off another half of a tenth of a second
when not finding duplicates. When finding duplicates this has no
practical affect as far as I can tell.

Also, return the average frequency (likelihood) across all candidates
in each cell as a list.