Solve more constrained cells first
Traverse the cells in order of greatest to fewest constraints.
This shaves another half of a tenth of a second off the solution when
checking for duplicates. When not checking for duplicates, it adds a
small amount of time, but this case is very fast anyway, the duplicate
check case always takes longer to run.